+++ 459/D.Pashmak_and_Parmida_problem.cpp
@@ -3,7 +3,7 @@
 #include<ext/pb_ds/tree_policy.hpp>
 using namespace __gnu_pbds;
 using namespace std;
-#define ll long long
+#define int long long
 const int N = 1e6 + 7;
 int l[N], r[N];
 template <typename T> using o_set = tree<T, null_type, less<T>, rb_tree_tag, tree_order_statistics_node_update>;
@@ -25,7 +25,7 @@ int main() {
   //for(int i = 1; i <= n; i++) cout << l[i] << " "; cout << "\n";
   //for(int i = 1; i <= n; i++) cout << r[i] << " "; cout << "\n";
   o_set<pair<int, int>> st;
-  ll ans = 0;
+  int ans = 0;
   for(int i = n; i >= 1; i--) {
     ans += st.order_of_key({l[i], i});
     st.insert({r[i], i});
